requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
python-multipart
python-dotenv
python-docx
